Small refactor and icons
